Minor fixes: remove a few useless includes and move a few others to the right file. Remove a few coverity warnings about not initialized members. Minor bug fix: Zone duplicate zone in legacy mode: duplicate a zone to the same layer (and therefore  the same location) is no more allowed.

+++ pcbnew/basepcbframe.cpp
@@ -850,7 +850,7 @@ void PCB_BASE_FRAME::updateZoomSelectBox()
 
 void PCB_BASE_FRAME::SetFastGrid1()
 {
-    if( m_FastGrid1 >= GetScreen()->GetGridCount() )
+    if( m_FastGrid1 >= (int)GetScreen()->GetGridCount() )
         return;
 
     int cmdId = GetScreen()->GetGrids()[m_FastGrid1].m_CmdId;
@@ -869,7 +869,7 @@ void PCB_BASE_FRAME::SetFastGrid1()
 
 void PCB_BASE_FRAME::SetFastGrid2()
 {
-    if( m_FastGrid2 >= GetScreen()->GetGridCount() )
+    if( m_FastGrid2 >= (int)GetScreen()->GetGridCount() )
         return;
 
     int cmdId = GetScreen()->GetGrids()[m_FastGrid2].m_CmdId;
